requirements.txt
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
idna==2.8
pysubs2==0.2.3
requests==2.21.0
urllib3==1.24.1
Whoosh==2.7.4
